add Assignment entity

+++ Gemstone.Infrastructure/DataInitialization/DbInitializer.cs
@@ -9,20 +9,29 @@ namespace Gemstone.Infrastructure.DataInitialization
 {
     public class DbInitializer
     {
-        private static Random _gen = new Random();
+        private static Random _random = new Random();
         private static DateTime _startDate = new DateTime(2000, 1, 1);
         private static int _range = 0;
+        private static bool _recreateDb = true;
 
         private static DateTime RandomDateTime()
         {
             _range = (DateTime.Today - _startDate).Days;
-            return _startDate.AddDays(_gen.Next(_range)).AddSeconds(_gen.Next(_range));
+            return _startDate.AddDays(_random.Next(_range)).AddSeconds(_random.Next(_range));
+        }
+
+        private static bool RandomBool()
+        {
+            return _random.NextDouble() > 0.5;
         }
 
         public static void SeedData(EfDbContext context)
         {
-            context.Database.EnsureDeleted();
-            context.Database.EnsureCreated();
+            if (_recreateDb)
+            {
+                context.Database.EnsureDeleted();
+                context.Database.EnsureCreated();
+            }
 
             if (context.Account.Any())
                 return;
@@ -36,17 +45,16 @@ namespace Gemstone.Infrastructure.DataInitialization
 
             var specialists = new Account[]
             {
-                new Specialist { Username = "Marcin" , AccountRole = AccountRole.Specialist, CraftAreaName = "motorcycle workshop/repair",Password = "specpass",    JoinedOn = RandomDateTime() },
-                new Specialist { Username = "Michal" , AccountRole = AccountRole.Specialist, CraftAreaName = "motorcycle workshop/repair",Password = "specpass",    JoinedOn = RandomDateTime() },
-                new Specialist { Username = "Pawel" ,AccountRole = AccountRole.Specialist, CraftAreaName = "motorcycle workshop/repair",Password = "specpass",      JoinedOn = RandomDateTime() },
-                new Specialist { Username = "Grzegorz" , AccountRole = AccountRole.Specialist, CraftAreaName = "motorcycle workshop/repair",Password = "specpass",  JoinedOn = RandomDateTime() },
-                new Specialist { Username = "Michal" , AccountRole = AccountRole.Specialist, CraftAreaName = "motorcycle workshop/repair",Password = "specpass",    JoinedOn = RandomDateTime() },
-                new Specialist { Username = "Jan" ,AccountRole = AccountRole.Specialist, CraftAreaName = "motorcycle workshop/repair",Password = "specpass",        JoinedOn = RandomDateTime() },
-                new Specialist { Username = "Jozef" , AccountRole = AccountRole.Specialist, CraftAreaName = "motorcycle workshop/repair",Password = "specpass",     JoinedOn = RandomDateTime() },
-                new Specialist { Username = "Marek" , AccountRole = AccountRole.Specialist, CraftAreaName = "motorcycle workshop/repair",Password = "specpass",     JoinedOn = RandomDateTime() },
-                new Specialist { Username = "Bogdan" ,AccountRole = AccountRole.Specialist, CraftAreaName = "motorcycle workshop/repair",Password = "specpass",     JoinedOn = RandomDateTime() },
+                new Specialist { Username = "Marcin" ,  Password = "specpass", IsBusy = RandomBool(),    AccountRole = AccountRole.Specialist, CraftAreaName = "motorcycle workshop/repair",    JoinedOn = RandomDateTime() },
+                new Specialist { Username = "Michal" ,  Password = "specpass", IsBusy = RandomBool(),    AccountRole = AccountRole.Specialist, CraftAreaName = "motorcycle workshop/repair",    JoinedOn = RandomDateTime() },
+                new Specialist { Username = "Pawel" ,   Password = "specpass", IsBusy = RandomBool(),    AccountRole = AccountRole.Specialist, CraftAreaName = "motorcycle workshop/repair",      JoinedOn = RandomDateTime() },
+                new Specialist { Username = "Grzegorz", Password = "specpass", IsBusy = RandomBool(),    AccountRole = AccountRole.Specialist, CraftAreaName = "motorcycle workshop/repair",  JoinedOn = RandomDateTime() },
+                new Specialist { Username = "Michal" ,  Password = "specpass", IsBusy = RandomBool(),    AccountRole = AccountRole.Specialist, CraftAreaName = "motorcycle workshop/repair",    JoinedOn = RandomDateTime() },
+                new Specialist { Username = "Jan" ,     Password = "specpass", IsBusy = RandomBool(),    AccountRole = AccountRole.Specialist, CraftAreaName = "motorcycle workshop/repair",        JoinedOn = RandomDateTime() },
+                new Specialist { Username = "Jozef" ,   Password = "specpass", IsBusy = RandomBool(),    AccountRole = AccountRole.Specialist, CraftAreaName = "motorcycle workshop/repair",     JoinedOn = RandomDateTime() },
+                new Specialist { Username = "Marek" ,   Password = "specpass", IsBusy = RandomBool(),    AccountRole = AccountRole.Specialist, CraftAreaName = "motorcycle workshop/repair",     JoinedOn = RandomDateTime() },
+                new Specialist { Username = "Bogdan" ,  Password = "specpass", IsBusy = RandomBool(),    AccountRole = AccountRole.Specialist, CraftAreaName = "motorcycle workshop/repair",     JoinedOn = RandomDateTime() },
             };
-
             foreach (var specialist in specialists)
                 context.Account.Add(specialist);
